feat: Added double-buffered stat accumulator and score-change telemetry
- Added StatAccumulatorService: reliable double-buffered JSON-over-TCP
  channel that flushes only when the buffer fills
- Reworked LiveStatsTracker to delegate buffering to the service
- Added ScoreChangeDto and subscribed to scoreDidChangeEvent
- Converted all song-time fields to integer milliseconds
- Logged received item counts at Information level in server handlers

+++ src/BeatDash.Data/Socket/MotionFrame.cs
@@ -15,12 +15,12 @@ public readonly struct TransformData {
 
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public readonly struct MotionFrame {
-    public readonly float SongTime;
+    public readonly int SongTime;
     public readonly TransformData LeftSaber;
     public readonly TransformData RightSaber;
     public readonly TransformData Head;
 
-    public MotionFrame(float songTime, TransformData leftSaber, TransformData rightSaber, TransformData head) {
+    public MotionFrame(int songTime, TransformData leftSaber, TransformData rightSaber, TransformData head) {
         SongTime = songTime;
         LeftSaber = leftSaber;
         RightSaber = rightSaber;
